Change synchronization for connected clients
Changes the way synchronous access to connected clients is handled. This is a
major change internally. The goal has been to refactor the way connected users
are handled to provide a easier API to work with internally. The new
implementation does make it easier to handle concurrency, however it does have
some caveats.

- Because each user has a mutex, it's easy to cause a deadlock in the functions
handling all the users in a chatroom. When you hold the lock to a single user
and iterate over all the users in the chatroom, the current user might be apart
of that iteration. This will cause a deadlock if you then lock the user being
iterated over.
- When a function locks a user, it must make sure that it does not call another
function which performs the same lookup. This should probably be refractored
so that every function which needs a user should have it provided while it's
already locked in the parent context.

The main change is the `Users` struct (name could be changed) which is contained withing
`server/user.go`. This struct defines a threadsafe API for accessing all the
connected clients. Each user is then protected by a mutex which MUST be held
for reading/writing to the user struct, however simple pointer comparisions can
be made without holding the mutex.

Additional changes:

- `CheckAuth` has been removed in favor of using `KeyMatches` directly on the
current user
- `FindClientsInChat` has been removed in favor of `ForEachInChat` on the
`Users` struct