Fix bug in remove client
Previously connected clients would not be removed if the user-object was nil.
This is a implementation bug because a nil user means that the websocket is not
yet associated with a user.

It will now always remove a websocket no matter if the associated user object
is set or not. Further it will log wether the connection was authenticated
and/or in a chatroom.